Move visibilityFilter to correct location
